Added test markers, moved function to pod.py
